Extract new <Badge /> element from <Cell />
[MAILPOET-5091]
